LayoutTests/fast/js/sputnik/Conformance/11_Expressions/11.1_Primary_Expressions/11.1.1_The_this_Keyword/S11.1.1_A1-expected.txt

S11.1.1_A1

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
